Create LiquiGuard: Anti-Liquidation Futures Trading Bot
We hate Liquidation when we open a position with high leverage. It may came a time you do not want to add extract margin. With this bot,you need not to worry about that. It would automatically open a reverse position when it is about to be liquidated and close it when it recovered based on well designed indicators(not only traditional ones).